+++ graph_query.h
